v0.0.2-beta 修改日线、复权的API，速度更快 当前 pandas 2.1 废弃了一些方法 与 clickhouse-concent 冲突，勿升级 修复了一次获取过多标的导致的bug

+++ script/init.sql
@@ -38,7 +38,7 @@ CREATE TABLE quant.ts_min
     `high` Decimal(9, 2),
     `low` Decimal(9, 2),
     `close` Decimal(9, 2),
-    `volume` UInt64,
+    `volume` Decimal(18, 2),
     `amount` Decimal(18, 2)
 )
 ENGINE = ReplacingMergeTree
